third_party/blink/web_tests/fast/css/invalidation/shadow-add-sheet-host-expected.txt

Insert a style element into a shadow tree affecting the host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(host2).color is "rgb(255, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(host2).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
